// src/app/core/models/transmittal.model.ts
export class Transmittal
// {
//     projectName: string;
//     transmittalNo: string;
//     subject?: string;
//     purpose?: string;
//     comments?: string;
//     creationDate?: Date;
//     createdBy?: string;
//     status?: string;
//     type?: string;
// }
{
    id?: 0;
    projectName: string;
    transmittalNo: string;
    subject: string;
    purpose: string;
    comments?: string;
    creationDate?: Date;
    ceatedBy?: string;
    status?: string;
    type?: string;
    // transmittalRecipients: [
    //   {
    //     id?: 0;
    //     gettRecName: string;
    //     gettRecEmail: string;
    //     recComp: string;
    //     recTitle: string;
    //     isAckRequired: true;
    //     isRespRequired: true;
    //     acknolgment: true;
    //     response: true;
    //     ackDate: Date;
    //     respDate: Date;
    //     ackRequiredDate: Date;
    //     respRequiredDate: Date;
    //     comments: string;
    //     transmittal: string
    //   }
    // ];
    // documents: [
    //   {
    //     id?: 0;
    //     documentsNumber: string;
    //     revision: string;
    //     revisionDate: Date;
    //     version: string;
    //     fileType: string;
    //     title: string;
    //     phase: string;
    //     classes: string;
    //     documentsFor: string;
    //     initiatedBy: string;
    //     disciplines: string;
    //     receiveDate: Date;
    //     replyRequired: string;
    //     replyRequiredBy: Date;
    //     repliedDate: Date;
    //     tqStatus: string;
    //     path: string;
    //     workflow: string;
    //     currentStep: string;
    //     ssize: string;
    //     isTransmitted: true;
    //     sstatus: string;
    //     confidential: true;
    //     additionalReference: string;
    //     reviewStatus: string;
    //     modelReference: string;
    //     createdBy: string;
    //     dateModified: Date;
    //     relatedItems: string;
    //     accessLevel: string;
    //     csiSpecCode: string;
    //     current: string;
    //     facilityCode: string;
    //     fileName: string;
    //     forecastSubmToClient: Date;
    //     jobNumber: string;
    //     lock: true;
    //     lastModifiedDate: Date;
    //     milestone: string;
    //     numberOfMarkups: 0;
    //     activityID: string;
    //     plannedSubmissionDate: Date;
    //     printSize: string;
    //     purchaseOrder: string;
    //     remarks: string;
    //     reviewSource: string;
    //     projectField1: string;
    //     projectField2: string;
    //     projectField3: string;
    //     projectField4: string;
    //     projectField5: string;
    //     projectField6: string;
    //     projectField7: string;
    //     filesType: {
    //       id: 0;
    //       name: string;
    //       code: string;
    //       description: string
    //     };
    //     documentsType: {
    //       id?: 0;
    //       name: string;
    //       code: string;
    //       description: string
    //     };
    //     documentsSubType: {
    //       id?: 0;
    //       name: string;
    //       code: string;
    //       description: string
    //     };
    //     discipline: {
    //       id?: 0;
    //       name: string;
    //       code: string;
    //       description: string
    //     };
    //     status: {
    //       id?: 0;
    //       name: string;
    //       code: string;
    //       description: string
    //     };
    //     orgnizer: {
    //       id?: 0;
    //       name: string;
    //       code: string;
    //       type: string;
    //       description: string
    //     };
    //     recipient: {
    //       id?: 0;
    //       name: string;
    //       code: string;
    //       type: string;
    //       description: string
    //     };
    //     accptanceCode: {
    //       id?: 0;
    //       number: 0;
    //       name: string
    //     };
    //     categories: {
    //       id?: 0;
    //       number: 0;
    //       name: string;
    //       targetAcceptNumber: 0;
    //       result: string
    //     };
    //     subSystem: {
    //       id?: 0;
    //       name: string;
    //       description: string
    //     };
    //     transmittals: [
    //       string
    //     ];
    //     hyberLinks: [
    //       {
    //         id?: 0;
    //         linkId: string;
    //         message: string;
    //         canView: true;
    //         canEdit: true;
    //         canDownload: true;
    //         createdBy: string;
    //         isPublic: string;
    //         creationDate: Date;
    //         expiryDate: Date;
    //         hyberLinkUsers: [
    //           {
    //             id?: 0;
    //             userName: string;
    //             userEmail: string;
    //             hyberLink: string
    //           }
    //         ];
    //         documents: [
    //           string
    //         ]
    //       }
    //     ]
    //   }
    // ]
  }